refactor getForeignName -> getPrettyName
