+++ features/records/clientHelpers.ts
@@ -1,3 +1,4 @@
+import { BasetoolRecord, PossibleRecordValues } from "@/features/records/types";
 import { Column } from "../fields/types"
 import { IFilter } from "../tables/types";
 import { isArray, isString } from "lodash";
@@ -23,4 +24,34 @@ export const filtersForHasMany = (
       verb: "and",
     },
   ];
+};
+
+/**
+ * This method tries to extract a pretty name from a record
+ */
+export const getForeignName = (
+  record: BasetoolRecord,
+  field?: string | undefined
+): string => {
+  // See if we have some `name` column set in the DB
+  let prettyName: PossibleRecordValues = "";
+
+  // Use the `nameColumn` attribute
+  if (field) {
+    prettyName = record[field];
+  } else {
+    // Try and find a common `name` columns
+    if (record.url) prettyName = record.url;
+    if (record.email) prettyName = record.email;
+    if (record.first_name) prettyName = record.first_name;
+    if (record.firstName) prettyName = record.firstName;
+    if (record.title) prettyName = record.title;
+    if (record.name) prettyName = record.name;
+  }
+
+  if (prettyName) return prettyName.toString();
+
+  if (record && record?.id) return record.id.toString();
+
+  return "";
 };
